Work in progress - Refactor autocomplete search box into seperate component and remove Downshift component for search suggestions
