+++ search-parts/src/models/ISuggestion.ts
@@ -7,4 +7,5 @@ export interface ISuggestion {
   icon?: string;
   groupName?: string;
   onSuggestionSelected?: (suggestion: ISuggestion) => void;
+  id?: string;
 }
